backend/requirements.txt
fastapi==0.115.12
pydantic-settings==2.8.1
langchain-community==0.3.20
langchain-google-genai==2.1.2
llm-guard==0.3.15
html-sanitizer==2.5.0
uvicorn==0.34.0
unstructured==0.17.2
markdown==3.7
sqlglot[rs]==26.12.0
faiss-cpu==1.10.0
numpy==2.2.4
sentence-transformers==4.0.1
alembic==1.15.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10